Adds of function and refactors test cases
