+++ packages/ae-observable/test/methods.spec.ts
@@ -1,88 +1,11 @@
+import createObservable, { of, Subscription } from '../src';
 
-import createObservable, { Subscription } from '../src';
-// import { resolveAfter } from '../src/fns';
-
-// @ts-ignore
-const Promise = window.Promise;
-
-describe('Observable', () => {
-  it('should subscribe to observable correctly', done => {
-    const obs = createObservable<number>((subscription: Subscription) => {
-      subscription.complete({ value: 'Hello' });
-    });
-
-    obs.subscribe({
-      onError: done,
-      onNext: () => done('Not called here'),
-      onComplete: (sub: Subscription, { value }) => {
-        expect(sub.isCancelled).toBe(false);
-        expect(value).toBe('Hello');
-        done();
-      },
-    });
-  });
-
-  it('should subscribe to streamed data correctly', done => {
-    const onNext = jest.fn();
-    const obs = createObservable<number>((subscription: Subscription) => {
-      subscription.next(1);
-      subscription.next(3);
-      subscription.complete();
-    });
-
-    obs.subscribe({
-      onError: done,
-      onNext,
-      onComplete: () => {
-        expect(onNext).toBeCalledTimes(2);
-        expect(onNext.mock.calls).toEqual([[1], [3]]);
-        done();
-      },
-    });
-  });
-
-  it('should go to stream error on throw error', done => {
-    const onError = jest.fn();
-    const onNext = jest.fn();
-
-    const obs = createObservable<number>((subscription: Subscription) => {
-      subscription.next(1);
-      subscription.throwError(5);
-      subscription.next(2);
-      subscription.throwError(6);
-      subscription.complete();
-    });
-
-    obs
-      .map(x => x + 10)
-      .subscribe({
-        onError,
-        onNext,
-        onComplete: () => {
-          expect(onNext).toBeCalledTimes(2);
-          expect(onNext).toHaveBeenCalledWith(11);
-          expect(onNext).toHaveBeenCalledWith(12);
-
-          expect(onError).toBeCalledTimes(2);
-          expect(onError).toHaveBeenCalledWith(5);
-          expect(onError).toHaveBeenCalledWith(6);
-          done();
-        },
-      });
-  });
-
+describe('Observable methods', () => {
   describe('Observable#map', () => {
-    
     it('should map over the items in the stream (increment)', done => {
       const onNext = jest.fn();
-      const obs = createObservable<number>((subscription: Subscription) => {
-        subscription.next(1);
-        subscription.next(3);
-        subscription.next(10);
-        subscription.complete();
-      });
 
-      obs
+      of(1, 3, 10)
         .map(x => x + 10)
         .subscribe({
           onError: done,
@@ -175,19 +98,13 @@ describe('Observable', () => {
     
     it('should chain the two in the stream (increment)', done => {
       const onNext = jest.fn();
-      const obs = createObservable<number>((subscription: Subscription) => {
-        subscription.next(1);
-        subscription.next(3);
-        subscription.next(10);
-        subscription.complete();
-      });
 
       const add10Stream = (x: number) => createObservable(sub => {
         sub.next(x + 10);
         sub.complete();
       });
 
-      obs
+      of(1, 3, 10)
         .chain(add10Stream)
         .subscribe({
           onError: done,
